Reuse api
