+++ eng/targets/Services.props
@@ -39,7 +39,6 @@
     <ServiceHubService Include="Microsoft.VisualStudio.LanguageServices.ProcessTelemetry" ClassName="Microsoft.CodeAnalysis.Remote.RemoteProcessTelemetryService+Factory" />
     <ServiceHubService Include="Microsoft.VisualStudio.LanguageServices.CompilationAvailable" ClassName="Microsoft.CodeAnalysis.Remote.RemoteCompilationAvailableService+Factory" />
     <ServiceHubService Include="Microsoft.VisualStudio.LanguageServices.StackTraceExplorer" ClassName="Microsoft.CodeAnalysis.Remote.RemoteStackTraceExplorerService+Factory" />
-    <ServiceHubService Include="Microsoft.VisualStudio.LanguageServices.SymbolTreeInfoIncrementalAnalyzer" ClassName="Microsoft.CodeAnalysis.Remote.RemoteSymbolTreeInfoIncrementalAnalyzer+Factory" />
   </ItemGroup>
 
   <!--
