+++ src/Workspaces/Remote/ServiceHub/Services/SymbolFinder/RemoteSymbolFinderService.cs
@@ -10,7 +10,9 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.FindSymbols;
+using Microsoft.CodeAnalysis.FindSymbols.SymbolTree;
 using Microsoft.CodeAnalysis.PooledObjects;
+using Microsoft.CodeAnalysis.Shared.Extensions;
 using Microsoft.CodeAnalysis.Shared.Utilities;
 using Microsoft.CodeAnalysis.Text;
 using Roslyn.Utilities;
@@ -170,6 +172,49 @@ namespace Microsoft.CodeAnalysis.Remote
 
                 return Convert(result, solution, cancellationToken);
             }, cancellationToken);
+        }
+
+        public ValueTask AnalyzeDocumentAsync(
+            Checksum solutionChecksum,
+            DocumentId documentId,
+            bool isMethodBodyEdit,
+            CancellationToken cancellationToken)
+        {
+            return RunServiceAsync(
+                solutionChecksum,
+                async solution =>
+                {
+                    var cacheService = solution.Services.GetRequiredService<SymbolTreeInfoCacheService>();
+                    await cacheService.AnalyzeDocumentAsync(solution.GetRequiredDocument(documentId), isMethodBodyEdit, cancellationToken).ConfigureAwait(false);
+                },
+                cancellationToken);
+        }
+
+        public ValueTask AnalyzeProjectAsync(
+            Checksum solutionChecksum,
+            ProjectId projectId,
+            CancellationToken cancellationToken)
+        {
+            return RunServiceAsync(
+                solutionChecksum,
+                async solution =>
+                {
+                    var cacheService = solution.Services.GetRequiredService<SymbolTreeInfoCacheService>();
+                    await cacheService.AnalyzeProjectAsync(solution.GetRequiredProject(projectId), cancellationToken).ConfigureAwait(false);
+                },
+                cancellationToken);
+        }
+
+        public ValueTask RemoveProjectAsync(ProjectId projectId, CancellationToken cancellationToken)
+        {
+            return RunServiceAsync(
+                cancellationToken =>
+                {
+                    var cacheService = GetWorkspaceServices().GetRequiredService<SymbolTreeInfoCacheService>();
+                    cacheService.RemoveProject(projectId);
+                    return default;
+                },
+                cancellationToken);
         }
 
         private sealed class FindLiteralReferencesProgressCallback : IStreamingFindLiteralReferencesProgress, IStreamingProgressTracker
